Add files via upload
added the following requirements: 
The last box is positioned in the bottom right corner of the window and stays there even when the window resizes.
Each box is 100x150 pixels (color: #eeeff2), with a 10px dotted line (color: #D0D0FF) on the left. Boxes are separated by 10 pixels of space.
When hovering over a box, the cursor changes to a hand and the box and font colors change (colors: yellow, goldenrod, respectively).
There are 10 pixels of space between the letters and the edge of the box.